fixed device scan

+++ internal/model/web/device_responses.go
@@ -2,7 +2,7 @@ package web
 
 type DeviceInputResponse struct {
 	PhLevel        float32 `json:"ph"`
-	Humidity       int     `json:"kelembapan"`
+	Humidity       float32 `json:"kelembapan"`
 	LightIntensity int     `json:"intensitasCahaya"`
 	SoilType       string  `json:"jenisTanah"`
 }
